STK/requirements.txt
PyQt5
requests
python-whois
